refactor(react-components): split EventInfo out of EventCard
